Move Store.Build to store.go; fetch.go is fetcher only
