+++ plugins/store.go
@@ -9,6 +9,8 @@ import (
 	"plugin"
 
 	"github.com/psyduck-etl/sdk"
+
+	"github.com/gastrodon/psyduck/parse"
 )
 
 // Store represents the .psyduck/ workspace directory: the single source of
@@ -56,6 +58,31 @@ func (s *Store) writeManifest(m map[string]string) error {
 		return err
 	}
 	return os.WriteFile(s.manifestPath(), b, 0o644)
+}
+
+// Build clones and compiles the declared plugins, writing the name → .so path
+// manifest to the store's manifest file. Used by the init command.
+func (s *Store) Build(specs []parse.PluginSpec) error {
+	if err := os.MkdirAll(s.pluginsDir(), os.ModeDir|os.ModePerm); err != nil {
+		return fmt.Errorf("failed to create plugins dir: %w", err)
+	}
+
+	tmpDir, err := os.MkdirTemp("", "psyduck-plugin-*")
+	if err != nil {
+		return fmt.Errorf("failed to make temp dir: %w", err)
+	}
+	f := &fetcher{store: s, tmpDir: tmpDir}
+	defer f.cleanup()
+
+	collected := make(map[string]string, len(specs))
+	for _, spec := range specs {
+		loc, err := f.fetch(spec)
+		if err != nil {
+			return fmt.Errorf("unable to fetch %s: %w", spec.Name, err)
+		}
+		collected[spec.Name] = loc
+	}
+	return s.writeManifest(collected)
 }
 
 // Load reads the manifest and opens every plugin listed in it.
